app/rasberry/requirements.txt
paho-mqtt==1.2.1
numpy==1.13.1
scipy==0.19.1
